WIP setup

+++ setup.sh
@@ -19,23 +19,27 @@ last_invalid
 sudo apt-get -y install zsh vim ctags curl
 last_invalid
 
-#zsh setup
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
-last_invalid
-ln -s $HOME/dotfiles/oyho.zsh-theme $HOME/.oh-my-zsh/themes/
-last_invalid
-mv $HOME/.zshrc $HOME/.zshrc_old
-last_invalid
-ln -s $HOME/dotfiles/debian_common.zshrc $HOME/.zshrc
-last_invalid
+#zsh and oh-my-zsh setup if not already done
+if ! [ -f "$HOME/.oh-my-zsh" ]; then
+  sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+  last_invalid
+  ln -s $HOME/dotfiles/oyho.zsh-theme $HOME/.oh-my-zsh/themes/
+  last_invalid
+  mv $HOME/.zshrc $HOME/.zshrc_old
+  last_invalid
+  ln -s $HOME/dotfiles/debian_common.zshrc $HOME/.zshrc
+  last_invalid
+fi
 
 #vim setup
-ln -s $HOME/dotfiles/.vimrc $HOME/.vimrc
-last_invalid
-git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
-last_invalid
-mkdir -p $HOME/.vim/swapfiles
-last_invalid
+if ! [ -f "$HOME/.vim/bundle/Vundle.vim" ]; then
+  git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+  last_invalid
+  ln -s $HOME/dotfiles/.vimrc $HOME/.vimrc
+  last_invalid
+  mkdir -p $HOME/.vim/swapfiles
+  last_invalid
+fi
 
 #Run PluginInstall in vim
 vim +PluginInstall +qall
